indexgatewaypb to logproto Signed-off-by: Owen Diehl <ow.diehl@gmail.com>
